+++ Example/AppDelegate.swift
@@ -4,6 +4,7 @@ import AVKit
 import GoogleMobileAds
 import AppCenter
 import AppCenterCrashes
+import RAnalytics
 
 @UIApplicationMain
 class AppDelegate: UIResponder, UIApplicationDelegate {
@@ -12,8 +13,9 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
 
     func application(_ application: UIApplication, didFinishLaunchingWithOptions launchOptions: [UIApplication.LaunchOptionsKey: Any]?) -> Bool {
         AppCenter.start(withAppSecret: Bundle.main.value(for: "AppCenterSecret"), services: [Crashes.self])
-        self.window?.tintColor = #colorLiteral(red: 0.7472071648, green: 0, blue: 0, alpha: 1)
+        self.window?.tintColor = UIColor(named: "Crimson")
         GADMobileAds.sharedInstance().start(completionHandler: nil)
+        AnalyticsManager.shared().set(loggingLevel: .debug)
         return true
     }