+++ Example/Controllers/ContactsListSettingsTableViewController.swift
@@ -8,7 +8,7 @@ protocol ContactsListDelegate: class {
 class ContactsListSettingsTableViewController: UITableViewController {
 
     var userContactList: [MAContact]? = []
-    weak var contactDelegate: ContactsListDelegate?
+    weak var delegate: ContactsListDelegate?
     var selectedContacts = [MAContact]()
     var allowMultipleSelection = true
 
@@ -22,12 +22,15 @@ class ContactsListSettingsTableViewController: UITableViewController {
         userContactList = getContactList()
         if userContactList?.count ?? 0 == 0 {
             userContactList = []
-            for autogen in 0...9 {
+            repeat {
+                let name = Self.randomFakeName()
                 userContactList?.append(
                     MAContact(id: UUID().uuidString,
-                              name: self.randomFakeName(),
-                              email: "name.\(autogen)@example.com"))
-            }
+                              name: name,
+                              email: Self.fakeMail(with: name)
+                    )
+                )
+            } while (userContactList?.count ?? 10) < 10
             updateContactList(list: self.userContactList)
         }
     }
@@ -36,14 +39,20 @@ class ContactsListSettingsTableViewController: UITableViewController {
         let cell = tableView.dequeueReusableCell(withIdentifier: "Cell", for: indexPath) as UITableViewCell? ?? UITableViewCell()
         if let userContactList = userContactList, userContactList.indices.contains(indexPath.row) {
             let contact = userContactList[indexPath.row]
-            cell.detailTextLabel?.text = "id: \(contact.id)"
-            cell.textLabel?.text = contact.name
-            if contactDelegate != nil {
-                if allowMultipleSelection {
-                    cell.accessoryType = selectedContacts.contains(contact) ? .checkmark : .none
-                } else {
-                    cell.accessoryType = .disclosureIndicator
-                }
+            cell.detailTextLabel?.numberOfLines = 3
+            let titleSize = cell.textLabel?.font.pointSize ?? 12
+            var names = contact.name?.components(separatedBy: " ")
+            cell.textLabel?.attributedText = NSMutableAttributedString()
+                .normal(names?.removeFirst() ?? "", fontSize: titleSize)
+                .bold(" " + (names?.joined() ?? ""), fontSize: titleSize)
+            let size = cell.detailTextLabel?.font.pointSize ?? 10
+            cell.detailTextLabel?.attributedText = NSMutableAttributedString()
+                .bold("\(NSLocalizedString("contact.id", comment: "")): ", fontSize: size)
+                .normal(contact.id, fontSize: size-3)
+                .bold("\n\(NSLocalizedString("contact.email", comment: "")): ", fontSize: size)
+                .normal(contact.email ?? "", fontSize: size)
+            if delegate != nil {
+                cell.accessoryType = selectedContacts.contains(contact) ? .checkmark : .none
             }
         }
         return cell
@@ -51,8 +60,14 @@ class ContactsListSettingsTableViewController: UITableViewController {
 
     override func tableView(_ tableView: UITableView, didSelectRowAt indexPath: IndexPath) {
         if let contact = userContactList?[indexPath.row] {
-
-            if allowMultipleSelection {
+            if delegate == nil {
+                editContact(title: "Edit Contact info",
+                            index: indexPath.row,
+                            contactId: contact.id,
+                            contactName: contact.name,
+                            contactEmail: contact.email)
+                tableView.deselectRow(at: indexPath, animated: true)
+            } else if allowMultipleSelection {
                 if selectedContacts.contains(contact) {
                     selectedContacts.removeAll { contactToRemove -> Bool in
                         contactToRemove == contact
@@ -62,15 +77,20 @@ class ContactsListSettingsTableViewController: UITableViewController {
                 }
                 tableView.reloadRows(at: [indexPath], with: .automatic)
             } else {
-                tableView.deselectRow(at: indexPath, animated: true)
-                selectedContacts = [contact]
+                if selectedContacts.first != contact {
+                    selectedContacts.removeAll()
+                    selectedContacts.append(contact)
+                } else {
+                    selectedContacts.removeAll()
+                }
+                tableView.reloadData()
             }
-            contactDelegate?.contactsController(self, didSelect: selectedContacts)
+            delegate?.contactsController(self, didSelect: selectedContacts)
         }
     }
 
     override func tableView(_ tableView: UITableView, didDeselectRowAt indexPath: IndexPath) {
-        if allowMultipleSelection, let selectedContact = userContactList?[indexPath.row] {
+        if let selectedContact = userContactList?[indexPath.row] {
             selectedContacts.removeAll(where: { contact in
                 contact == selectedContact
             })
@@ -83,7 +103,7 @@ class ContactsListSettingsTableViewController: UITableViewController {
     }
 
     override func tableView(_ tableView: UITableView, canEditRowAt indexPath: IndexPath) -> Bool {
-        return contactDelegate == nil
+        return delegate == nil
     }
 
     override func tableView(_ tableView: UITableView, commit editingStyle: UITableViewCell.EditingStyle, forRowAt indexPath: IndexPath) {
@@ -97,44 +117,41 @@ class ContactsListSettingsTableViewController: UITableViewController {
     }
 
     @IBAction func addContact() {
-        self.addCustomContactId(title: "Please enter the custom ID you would like to add in Contacts", message: "")
+        self.addCustomContactId(title: "Please enter custom contact details", message: "")
     }
 
     func addCustomContactId(title: String, message: String) {
+        let contactName = Self.randomFakeName()
         DispatchQueue.main.async {
-            self.getInputFromAlertWithTextField(title: title, message: message, textFieldDefaultValue: UUID().uuidString) { (_, textField) in
-                if let textField = textField, let contactId = textField.text, contactId.count > 0, !contactId.trimTrailingWhitespaces().isEmpty {
-                    let contactListCount = self.userContactList?.count ?? 0
-                    self.userContactList?.append(MAContact(id: contactId,
-                                                           name: self.randomFakeName(),
-                                                           email: "name.\(contactListCount)@example.com"))
-                    updateContactList(list: self.userContactList)
-                    self.tableView.reloadData()
-                } else {
-                    self.addCustomContactId(title: "Invalid Contact ID, please try again", message: "Enter valid contact and select Ok")
-                }
-            }
+            self.editContact(title: title, index: 0, message: message, contactId: UUID().uuidString, contactName: contactName, contactEmail: Self.fakeMail(with: contactName), isNewContact: true)
         }
     }
 
-    func getInputFromAlertWithTextField(title: String? = nil, message: String? = nil, keyboardType: UIKeyboardType? = .asciiCapable, textFieldDefaultValue: String?, handler: ((UIAlertAction, UITextField?) -> Void)? = nil) {
+    func getInputFromAlertWithTextField(title: String? = nil,
+                                        message: String? = nil,
+                                        keyboardType: UIKeyboardType? = .asciiCapable,
+                                        textFieldDefaultValue: String?,
+                                        name: String? = randomFakeName(),
+                                        email: String? = nil,
+                                        handler: ((UIAlertAction, UITextField?, String, String) -> Void)? = nil) {
         DispatchQueue.main.asyncAfter(deadline: .now()) {
             let alert = UIAlertController(title: title,
                 message: message,
                 preferredStyle: .alert)
             var textObserver: NSObjectProtocol?
 
+            let email = email ?? Self.fakeMail(with: name)
+
             let okAction = UIAlertAction(title: MASDKLocale.localize(.ok), style: .default) { (action) in
                 if !alert.textFields![0].text!.isEmpty {
-                    handler?(action, alert.textFields?.first)
+                    handler?(action, alert.textFields?.first, alert.textFields?[1].text ?? "", alert.textFields?[2].text ?? "")
                 } else {
-                    handler?(action, nil)
+                    handler?(action, nil, "", "")
                 }
                 if let observer = textObserver {
                     NotificationCenter.default.removeObserver(observer)
                 }
             }
-
             alert.addTextField { (textField) in
                 textObserver = NotificationCenter.default.addObserver(forName: UITextField.textDidChangeNotification, object: textField, queue: OperationQueue.main) {_ in
                     okAction.isEnabled = !(textField.text?.isEmpty ?? true)
@@ -143,9 +160,20 @@ class ContactsListSettingsTableViewController: UITableViewController {
                 if let type = keyboardType {
                     textField.keyboardType = type
                 }
+                textField.placeholder = NSLocalizedString("contact.id", comment: "")
                 textField.clearButtonMode = .whileEditing
             }
 
+            alert.addTextField { (textField) in
+                textField.text = name
+                textField.placeholder = NSLocalizedString("contact.name", comment: "")
+                textField.clearButtonMode = .whileEditing
+            }
+            alert.addTextField { (textField) in
+                textField.text = email
+                textField.placeholder = NSLocalizedString("contact.email", comment: "")
+                textField.clearButtonMode = .whileEditing
+            }
             okAction.isEnabled = !(textFieldDefaultValue?.isEmpty ?? true)
             alert.addAction(UIAlertAction(title: MASDKLocale.localize(.cancel), style: .cancel, handler: nil))
             alert.addAction(okAction)
@@ -153,17 +181,67 @@ class ContactsListSettingsTableViewController: UITableViewController {
         }
     }
 
-    func randomFakeName() -> String {
+    func editContact(title: String, index: Int, message: String? = "", contactId: String? = "", contactName: String? = "", contactEmail: String? = "", isNewContact: Bool? = false) {
+        DispatchQueue.main.async {
+            self.getInputFromAlertWithTextField(title: title, message: message, textFieldDefaultValue: contactId, name: contactName, email: contactEmail) { (_, textField, name, email) in
+                self.validateAllValues(index: index, contactId: contactId, textField: textField, name: name, email: email, isNewContact: isNewContact)
+            }
+        }
+    }
+
+    func validateAllValues(index: Int, contactId: String?, textField: UITextField?, name: String, email: String, isNewContact: Bool? = false) {
+        if let contactIdTextField = textField {
+            if contactIdTextField.isTextFieldEmpty() {
+                self.editContact(title: "Invalid Contact ID, please try again",
+                                 index: index,
+                                 contactId: contactIdTextField.text,
+                                 contactName: name,
+                                 contactEmail: email, isNewContact: isNewContact)
+            } else if name.isValueEmpty() {
+                self.editContact(title: "Invalid Name, please try again",
+                                 index: index,
+                                 contactId: contactIdTextField.text,
+                                 contactName: name,
+                                 contactEmail: email, isNewContact: isNewContact)
+            } else if email.isValueEmpty() || !email.isValidEmail() {
+                self.editContact(title: "Invalid Email ID, please try again",
+                                 index: index,
+                                 contactId: contactIdTextField.text,
+                                 contactName: name,
+                                 contactEmail: email, isNewContact: isNewContact)
+            } else {
+                if let contactID =  textField?.text {
+                    if isNewContact ?? false {
+                        self.userContactList?.append(MAContact(id: contactID,
+                                                               name: name,
+                                                               email: email))
+                        updateContactList(list: self.userContactList)
+                        self.tableView.reloadData()
+                    } else {
+                        self.userContactList?[index] = MAContact(id: contactID, name: name, email: email)
+                        updateContactList(list: self.userContactList)
+                        self.tableView.reloadData()
+                    }
+                }
+            }
+        }
+    }
+
+    class func fakeMail(with name: String?) -> String {
+        name != nil ? name!.replacingOccurrences(of: " ", with: ".", options: .literal, range: nil).lowercased() + "@example.com" : ""
+    }
+
+    class func randomFakeName() -> String {
         randomFakeFirstName() + " " + randomFakeLastName()
     }
 
-    func randomFakeFirstName() -> String {
-        let firstNameList = ["哲也", "太郎", "ピエール", "レオ", "Yvonne", "Jamie", "Leticia", "Priscilla", "Sidney", "Nancy", "Edmund", "Bill", "Megan"]
+    class func randomFakeFirstName() -> String {
+        let firstNameList = ["Yvonne", "Jamie", "Leticia", "Priscilla", "Sidney", "Nancy", "Edmund", "Bill", "Megan"]
         return firstNameList.randomElement()!
     }
 
-    func randomFakeLastName() -> String {
-        let lastNameList = ["古室", "楽天", "ビラ", "ジョゼフ", "Andrews", "Casey", "Gross", "Lane", "Thomas", "Patrick", "Strickland", "Nicolas", "Freeman"]
+    class func randomFakeLastName() -> String {
+        let lastNameList = ["Andrews", "Casey", "Gross", "Lane", "Thomas", "Patrick", "Strickland", "Nicolas", "Freeman"]
         return lastNameList.randomElement()!
     }
 }