+++ Example/Controllers/Extensions/ViewController+MiniAppUserInfoProtocol.swift
@@ -16,16 +16,25 @@ extension ViewController: MiniAppUserInfoDelegate {
         completionHandler(.success(userProfilePhoto))
     }
 
-    func getContacts() -> [MAContact]? {
-        guard let userProfile = getProfileSettings(), let contactList = userProfile.contactList else {
-            return nil
+    func getContacts(completionHandler: @escaping (Result<[MAContact]?, MASDKError>) -> Void) {
+        if let userProfile = getProfileSettings(), let contactsList = userProfile.contactList {
+            return completionHandler(.success(contactsList))
         }
-        return contactList
+        completionHandler(.failure(.unknownError(domain: "Unknown Error", code: 1, description: "Failed to retrieve contacts list")))
     }
 
     func getAccessToken(miniAppId: String,
                         scopes: MASDKAccessTokenScopes,
-                        completionHandler: @escaping (Result<MATokenInfo, MASDKCustomPermissionError>) -> Void) {
+                        completionHandler: @escaping (Result<MATokenInfo, MASDKAccessTokenError>) -> Void) {
+        if let errorMode = QASettingsTableViewController.accessTokenErrorType() {
+            let message = QASettingsTableViewController.accessTokenErrorMessage()
+            switch errorMode {
+            case .AUTHORIZATION:
+                return completionHandler(.failure(.authorizationFailureError(description: "authorizationFailureError" + ( (message != nil) ? ": " + message! : "" ))))
+            default:
+                return completionHandler(.failure(.error(description: "Other error" + ( (message != nil) ? ": " + message! : "" ))))
+            }
+        }
         var resultToken = "ACCESS_TOKEN"
         var resultDate = Date()
         let resultScopes = scopes