+++ MiniApp/Classes/core/Downloader/MiniAppDownloader.swift
@@ -29,14 +29,12 @@ class MiniAppDownloader {
     }
 
     fileprivate func cleanApp(_ appId: String, for version: String) {
-        if self.miniAppClient.environment.isPreviewMode == false {
-            self.miniAppStorage.cleanVersions(for: appId, differentFrom: version, status: self.miniAppStatus)
-        }
+        miniAppStorage.cleanVersions(for: appId, differentFrom: version, status: miniAppStatus)
     }
 
     func verifyAndDownload(appId: String, versionId: String, completionHandler: @escaping (Result<URL, Error>) -> Void) {
         if isMiniAppAlreadyDownloaded(appId: appId, versionId: versionId) {
-            if cacheVerifier.verify(appId: appId, version: versionId) {
+            if !miniAppClient.environment.isPreviewMode, cacheVerifier.verify(appId: appId, version: versionId) {
                 cleanApp(appId, for: versionId)
                 let miniAppStoragePath = FileManager.getMiniAppVersionDirectory(with: appId, and: versionId)
                 MiniAppLogger.d("\(miniAppStoragePath.absoluteString)", "📂")