+++ MiniApp/Classes/core/Protocols/MiniAppMessageDelegate.swift
@@ -51,7 +51,19 @@ public extension MiniAppMessageDelegate {
     @available(*, deprecated,
         renamed: "getUniqueId(completionHandler:)")
     func getUniqueId() -> String? {
-        return nil
+        let semaphore = DispatchSemaphore(value: 0)
+        var uniqueId: String?
+        getUniqueId { result in
+            switch result {
+            case .success(let id):
+                uniqueId = id
+            default:
+                uniqueId = nil
+            }
+            semaphore.signal()
+        }
+        semaphore.wait()
+        return uniqueId
     }
 }
 
@@ -63,78 +75,4 @@ public enum MASDKProtocolResponse: String {
 public enum MASDKPermissionResponse: String {
     /// User allowed the Device Permission
     case allowed = "ALLOWED"
-}
-
-/// Enumeration that is used to differentiate the device permission errors
-public enum MASDKPermissionError: String, MiniAppErrorProtocol {
-
-    /// User has explicitly denied authorization
-    case denied = "DENIED"
-    /// User has not yet made a choice
-    case notDetermined = "NOT_DETERMINED"
-    /// Host app is not authorized to use location services
-    case restricted = "RESTRICTED"
-    /// Host app failed to implement required interface
-    case failedToConformToProtocol = "FAILED_TO_CONFORM_PROTOCOL"
-
-    var name: String {
-        return self.rawValue
-    }
-
-    var description: String {
-        switch self {
-        case .denied:
-            return "User has explicitly denied authorization"
-        case .notDetermined:
-            return "User has not yet made a choice"
-        case .restricted:
-            return "Host app is not authorized to use location services"
-        case .failedToConformToProtocol:
-            return "Host app failed to implement required interface"
-        }
-    }
-}
-
-/// Enumeration that is used to differentiate the Custom permission errors
-public enum MASDKCustomPermissionError: String, MiniAppErrorProtocol {
-
-    /// Unknown Error
-    case unknownError = "UKNOWN_ERROR"
-
-    /// Host app failed to implement required interface
-    case failedToConformToProtocol = "FAILED_TO_CONFORM_PROTOCOL"
-
-    /// Invalid Custom Permission request from Mini app
-    case invalidCustomPermissionRequest
-
-    /// Invalid list of Custom Permission requested from Mini app
-    case invalidCustomPermissionsList
-
-    /// User denied the Custom Permission
-    case userDenied
-
-    /// Invalid scope request for the Custom Permission
-    case outOfScope
-
-    var name: String {
-        return self.rawValue
-    }
-
-    /// Detailed Description for every MASDKCustomPermissionError
-    public var description: String {
-        switch self {
-        case .unknownError:
-            return "Unknown error occurred"
-        case .failedToConformToProtocol:
-            return "Host app failed to implement required interface"
-        case .invalidCustomPermissionRequest:
-            return "Error in Custom Permission Request, please make sure the Custom permissions are passed in []"
-        case .invalidCustomPermissionsList:
-            return "Error in list of Custom permissions that is passed, please check whether valid permission associated with name "
-        case .userDenied:
-            return "User denied to share the detail"
-        case .outOfScope:
-            return "Invalid scope request for the Custom Permission"
-        }
-    }
 }